backend/requirements.txt
fastapi==0.114.2
pydantic==2.9.1
uvicorn==0.30.6
python-dotenv==1.0.1
gunicorn>=21.2.0
pyodbc
azure-identity
